Поддержка Теней Чернобыля
- [xrGameSoC]: Правки под VS2019
- [xrGameSoC]: Поддержка x64
- Упрощение кода и ввод чтения нового режима xrGame – SoC
- Отдельная загрузка нужных параметров партиклов в зависимости от xrGame
- Необходимые для запуска текстуры и конфиги
- Фиксы синтаксиса в конфигах
- Чтение textures.ltx
- OnScreenRatioChanged -> OnScreenResolutionChanged
- Добавлены шейдеры Advanced X-Ray
- Добавлен правленный shaders.xr под ТЧ
- Чтение лайтмапов Теней Чернобыля (Автор: @SkyLoaderr)
- Опциональный парсинг дескрипшенов текстур из LA

- Правки загрузочного экрана:

1. Автопауза из ЗП
2. Окно загрузки выведено в конфиги (OpenXRay)
3. Добавлен интерфейс UIHelper из ЧН
4. Добавлены опциональные советы по выживанию в Зоне
5. Синхронизация кода стадий загрузки игры
6. Исправления от @Hrusteckiy

Консольная команда: keypress_on_start 0-1 - вкл.\выкл. автопаузы.

- Запуск игры

+++ SourcesAXR/xrGameSoC/sound_memory_manager.cpp
@@ -23,7 +23,7 @@
 #include "profiler.h"
 #include "client_spawn_manager.h"
 #include "memory_manager.h"
-#include "../IGame_Persistent.h"
+#include "../xrEngine/igame_persistent.h"
 
 #ifndef MASTER_GOLD
 #	include "clsid_game.h"
@@ -91,7 +91,7 @@ IC	void CSoundMemoryManager::update_sound_threshold			()
 	m_sound_threshold		= _max(
 		m_self_sound_factor*
 		m_sound_threshold*
-		exp(
+		(float)exp(
 			float(Device.dwTimeGlobal - m_last_sound_time)/
 			float(m_sound_decrease_quant)*
 			log(m_decrease_factor)
